refactoring(product) fix product colection module relations
