+++ module/product-collection/src/Infrastructure/Persistence/Query/DbalProductCollectionQuery.php
@@ -18,10 +18,12 @@ use Ergonode\ProductCollection\Domain\Query\ProductCollectionQueryInterface;
 use Ergonode\ProductCollection\Domain\ValueObject\ProductCollectionCode;
 use Ergonode\SharedKernel\Domain\Aggregate\ProductCollectionId;
 use Ergonode\SharedKernel\Domain\Aggregate\ProductCollectionTypeId;
+use Ergonode\SharedKernel\Domain\Aggregate\ProductId;
 
 class DbalProductCollectionQuery implements ProductCollectionQueryInterface
 {
     private const PRODUCT_COLLECTION_TABLE = 'public.product_collection';
+    private const PRODUCT_COLLECTION_ELEMENT_TABLE = 'public.product_collection_element';
 
     private Connection $connection;
 
@@ -78,6 +80,33 @@ class DbalProductCollectionQuery implements ProductCollectionQueryInterface
         return $this->dataSetFactory->create($result);
     }
 
+    public function getDataSetByProduct(Language $language, ProductId $productId): DataSetInterface
+    {
+        $qb = $this->connection->createQueryBuilder();
+        $qb->select('id, code, type_id')
+            ->addSelect(sprintf('(name->>\'%s\') AS name', $language->getCode()))
+            ->addSelect(sprintf('(description->>\'%s\') AS description', $language->getCode()))
+            ->addSelect('(SELECT count(*) FROM product_collection_element'.
+            ' WHERE product_collection_id = c.id) as elements_count')
+            ->from(self::PRODUCT_COLLECTION_TABLE, 'c')
+            ->leftJoin(
+                'c',
+                self::PRODUCT_COLLECTION_ELEMENT_TABLE,
+                'ce',
+                'ce.product_collection_id = c.id'
+            )
+            ->where($qb->expr()->eq('product_id', ':product_id'))
+            ->andWhere('visible=true');
+
+        $result = $this->connection->createQueryBuilder();
+
+        $result->setParameter(':product_id', $productId->getValue());
+        $result->select('*');
+        $result->from(sprintf('(%s)', $qb->getSQL()), 't');
+
+        return $this->dataSetFactory->create($result);
+    }
+
     public function findIdByCode(ProductCollectionCode $code): ?ProductCollectionId
     {
         $qb = $this->connection->createQueryBuilder();
@@ -106,6 +135,30 @@ class DbalProductCollectionQuery implements ProductCollectionQueryInterface
 
         $result = $qb
             ->where($qb->expr()->eq('type_id', ':id'))
+            ->setParameter(':id', $id->getValue())
+            ->execute()
+            ->fetchAll(\PDO::FETCH_COLUMN);
+
+        if (false === $result) {
+            $result = [];
+        }
+
+        foreach ($result as &$item) {
+            $item = new ProductCollectionId($item);
+        }
+
+        return $result;
+    }
+
+    /**
+     * @return mixed|void
+     */
+    public function findProductCollectionIdByProductId(ProductId $id)
+    {
+        $qb = $this->connection->createQueryBuilder();
+        $result = $qb->select('pc.product_collection_id')
+            ->from(self::PRODUCT_COLLECTION_ELEMENT_TABLE, 'pc')
+            ->where($qb->expr()->eq('product_id', ':id'))
             ->setParameter(':id', $id->getValue())
             ->execute()
             ->fetchAll(\PDO::FETCH_COLUMN);
