+++ module/product-collection/src/Infrastructure/Relationship/ProductProductCollectionRelationshipStrategy.php
@@ -7,37 +7,31 @@
 
 declare(strict_types=1);
 
-namespace Ergonode\Product\Infrastructure\Strategy\Relationship;
+namespace Ergonode\ProductCollection\Infrastructure\Relationship;
 
 use Ergonode\Core\Infrastructure\Strategy\RelationshipStrategyInterface;
-use Ergonode\Product\Domain\Query\ProductQueryInterface;
 use Ergonode\SharedKernel\Domain\Aggregate\ProductId;
 use Ergonode\SharedKernel\Domain\AggregateId;
 use Webmozart\Assert\Assert;
 use Ergonode\Core\Infrastructure\Model\RelationshipGroup;
+use Ergonode\ProductCollection\Domain\Query\ProductCollectionQueryInterface;
 
 class ProductProductCollectionRelationshipStrategy implements RelationshipStrategyInterface
 {
     private const MESSAGE = 'Object has active relationships with {relations}';
 
-    private ProductQueryInterface $query;
+    private ProductCollectionQueryInterface $query;
 
-    public function __construct(ProductQueryInterface $query)
+    public function __construct(ProductCollectionQueryInterface $query)
     {
         $this->query = $query;
     }
 
-    /**
-     * {@inheritDoc}
-     */
     public function supports(AggregateId $id): bool
     {
         return $id instanceof ProductId;
     }
 
-    /**
-     * {@inheritDoc}
-     */
     public function getRelationshipGroup(AggregateId $id): RelationshipGroup
     {
         Assert::isInstanceOf($id, ProductId::class);
